Move WebcilConverter to Microsoft.NET.WebAssembly.Webcil
