+++ src/libraries/Microsoft.NET.WebAssembly.Webcil/Directory.Build.props
@@ -5,5 +5,7 @@
       <!-- this assembly should not produce a public package, rather it's meant to be shipped by the
            WasmAppBuilder task and the BrowserDebugProxy -->
       <IsShippingPackage>false</IsShippingPackage>
+      <!-- This isn't a public API in a public package, don't ship documentation xml in the nugets that consume this assembly -->
+      <GenerateDocumentationFile>false</GenerateDocumentationFile>
   </PropertyGroup>
 </Project>
