+++ src/libraries/Microsoft.NET.WebAssembly.Webcil/src/Webcil/Internal/Constants.cs
@@ -1,9 +1,9 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-namespace Microsoft.WebAssembly.Build.Tasks.WebCil;
+namespace Microsoft.NET.WebAssembly.Webcil.Internal;
 
-public static unsafe class Constants
+internal static unsafe class Constants
 {
     public const int WC_VERSION = 0;
 }
